+++ src_test/tbrugz/sqldiff/SQLDiffMainTest.java
@@ -3,6 +3,7 @@ package tbrugz.sqldiff;
 import java.io.IOException;
 import java.sql.SQLException;
 import java.util.Properties;
+import java.util.concurrent.ExecutionException;
 
 import javax.naming.NamingException;
 import javax.xml.bind.JAXBException;
@@ -18,7 +19,7 @@ public class SQLDiffMainTest {
 	String OUTDIR = "work/output/SQLDiffMainTest";
 	
 	@Test
-	public void testMain1() throws IOException, ClassNotFoundException, SQLException, NamingException, JAXBException, XMLStreamException {
+	public void testMain1() throws IOException, ClassNotFoundException, SQLException, NamingException, JAXBException, XMLStreamException, InterruptedException, ExecutionException {
 		Properties p = new ParametrizedProperties();
 		p.load(SQLDiffMainTest.class.getResourceAsStream("diff1.properties"));
 		p.setProperty("outputdir", OUTDIR);